Move ProgressBlobs to shared components
